+++ CHANGELOG.rst
@@ -1,12 +1,14 @@
 1.1.0 - 
 ------------------
 
-* added Gtk version (selected with --gtk2 flag)
+* added Gtk version (selected with ``--gtk2`` flag)
 
 * Tkinter now version works in Python 3.
 
 * switched default shell to ``/bin/sh``. (Not everyone has
   ``/bin/bash`` installed.)
+
+* added ``--shell`` option. Overrides ``$SHELL`` variable.
 
 * added proper argument parsing.
 
@@ -21,8 +23,8 @@
 1.0.1 - 2014-09-27
 ------------------
 
-* changed the WM_CLASS from the default "Tk" to "Launchbox" to make it
-  easy to make the window floating in XMonad (and other tiling
+* changed the ``WM_CLASS`` from the default "Tk" to "Launchbox" to
+  make it easy to make the window floating in XMonad (and other tiling
   wm:s). (Patch by Jonas Juselius.)